refactor(components): convert provider function to index.ts
